requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
bcrypt
werkzeug
psycopg2-binary
